Add files via upload
The purpose of this project is to build an API in Python that fetches responses from ChatGPT, then stores them in a SQL database. The user input's a question and my app collects responses from ChatGPT before displaying them on a custom built UI. Then it will automatically generate follow-up questions related to the original response allowing the user to click and find out more. Additionally, all responses and questions would be added to a conversation log allowing the user to re-read any response or follow up question at a later time. I have included comments to define each lines purpose and what it does in my code.